refactor(datastorage): split audit_events_repository.go by concern (Wave 3)
Split pkg/datastorage/repository/audit_events_repository.go (1,204 lines)
into audit_events_repository.go (trimmed: AuditEvent model + constructor),
audit_events_hashchain.go (hash-chain algorithm + verification),
audit_events_create.go (Create), audit_events_batch.go (CreateBatch), and
audit_events_query.go (Query + row scanning + HealthCheck).

Pure code motion, no behavior change, no exported-API change. See audit
report GO-ANTIPATTERN-AUDIT-2026-07-01 §7k-3.
