+++ pkg/datastorage/repository/audit_events_repository.go
@@ -17,22 +17,12 @@ limitations under the License.
 package repository
 
 import (
-	"context"
-	"crypto/hmac"
-	"crypto/sha256"
 	"database/sql"
-	"encoding/hex"
-	"encoding/json"
-	"errors"
 	"fmt"
-	"sort"
 	"time"
 
 	"github.com/go-logr/logr"
 	"github.com/google/uuid"
-
-	"github.com/jordigilh/kubernaut/pkg/datastorage/repository/sqlutil"
-	"github.com/jordigilh/kubernaut/pkg/datastorage/repository/txretry"
 )
 
 // DateOnly is a time.Time that serializes to JSON as date-only format (YYYY-MM-DD)
@@ -115,6 +105,16 @@ func (d DateOnly) Value() (interface{}, error) {
 // - Event sourcing pattern (immutable, append-only)
 // - Monthly range partitioning
 // - JSONB hybrid storage (26 structured columns + flexible event_data)
+//
+// File layout (GO-ANTIPATTERN-AUDIT-2026-07-01 Wave 3 split, pure code
+// motion, no behavior change):
+//   - audit_events_repository.go (this file): AuditEvent model, DateOnly,
+//     AuditEventsRepository struct + constructor
+//   - audit_events_hashchain.go: SOC2 Gap #9 hash-chain primitives
+//   - audit_events_create.go: single-event Create
+//   - audit_events_batch.go: multi-event CreateBatch
+//   - audit_events_query.go: Query, row-scanning, HealthCheck
+//   - audit_export.go: Export (CSV/JSON audit export)
 //
 // ========================================
 
@@ -257,844 +257,4 @@ func NewAuditEventsRepository(db *sql.DB, logger logr.Logger, opts ...AuditEvent
 // using the same key the repository uses at write time.
 func (r *AuditEventsRepository) HMACKey() []byte {
 	return r.hmacKey
-}
-
-// ========================================
-// GAP #9: HASH CHAIN IMPLEMENTATION (Tamper-Evidence)
-// Authority: AUDIT_V1_0_ENTERPRISE_COMPLIANCE_PLAN_DEC_18_2025.md - Day 7
-// SOC2 Requirement: Tamper-evident audit logs (SOC 2 Type II, NIST 800-53, Sarbanes-Oxley)
-// ========================================
-
-// GAP-05 (Issue #1505): Hash chain algorithm identifiers, stored per-row in the
-// hash_algorithm column so a mixed-algorithm chain is supported during HMAC key
-// rollout (existing rows keep sha256-unkeyed; new writes use hmac-sha256 once a
-// key is configured).
-const (
-	// HashAlgorithmSHA256Unkeyed is the legacy algorithm used before GAP-05: a
-	// plain SHA256 of (previous_hash + event_json), computable by anyone without
-	// a secret. A DB-privileged attacker can recompute this hash after tampering.
-	HashAlgorithmSHA256Unkeyed = "sha256-unkeyed"
-
-	// HashAlgorithmHMACSHA256 is the GAP-05 keyed algorithm: HMAC-SHA256 of
-	// (previous_hash + event_json) using a key stored outside the database
-	// (Kubernetes Secret). Forging a valid hash without the key is infeasible.
-	HashAlgorithmHMACSHA256 = "hmac-sha256"
-)
-
-// PrepareEventForHashing returns a copy of the event with excluded fields zeroed out.
-// This MUST be used by all hash calculation paths (write-time, export, verify-chain)
-// to ensure consistent hashing across the entire audit pipeline.
-//
-// Excluded fields:
-//  1. EventHash, PreviousEventHash — not yet calculated at write time
-//  2. EventDate — derived from EventTimestamp (DB-generated)
-//  3. LegalHold, LegalHoldReason, LegalHoldPlacedBy, LegalHoldPlacedAt — can change
-//     after event creation (SOC2 Gap #8)
-//  4. HashAlgorithm — GAP-05 metadata describing which algorithm produced EventHash,
-//     not part of the hashed content. Excluding it keeps pre-GAP-05 hashes verifiable
-//     (their original JSON payload never contained this field).
-//
-// Note: EventTimestamp IS included in hash (set before calculation during INSERT).
-func PrepareEventForHashing(event *AuditEvent) AuditEvent {
-	eventCopy := *event
-	eventCopy.EventHash = ""
-	eventCopy.PreviousEventHash = ""
-	eventCopy.EventDate = DateOnly{}
-	eventCopy.HashAlgorithm = ""
-
-	// SOC2 Gap #8: Legal hold fields can change after event creation
-	eventCopy.LegalHold = false
-	eventCopy.LegalHoldReason = ""
-	eventCopy.LegalHoldPlacedBy = ""
-	eventCopy.LegalHoldPlacedAt = nil
-
-	return eventCopy
-}
-
-// calculateEventHash computes SHA256 hash for blockchain-style chain
-// Hash = SHA256(previous_event_hash + event_json)
-// This creates an immutable chain where tampering with ANY event breaks the chain
-//
-// GAP-05 (Issue #1505): kept for the legacy HashAlgorithmSHA256Unkeyed algorithm.
-// New writes prefer calculateEventHashHMAC when a key is configured (see
-// AuditEventsRepository.hashEvent).
-func calculateEventHash(previousHash string, event *AuditEvent) (string, error) {
-	eventForHashing := PrepareEventForHashing(event)
-
-	// Serialize event to JSON (canonical form for consistent hashing)
-	eventJSON, err := json.Marshal(eventForHashing)
-	if err != nil {
-		return "", fmt.Errorf("failed to marshal event for hashing: %w", err)
-	}
-
-	// Compute hash: SHA256(previous_hash + event_json)
-	hasher := sha256.New()
-	hasher.Write([]byte(previousHash))
-	hasher.Write(eventJSON)
-	hashBytes := hasher.Sum(nil)
-
-	return hex.EncodeToString(hashBytes), nil
-}
-
-// calculateEventHashHMAC computes a keyed HMAC-SHA256 hash chain link.
-// Hash = HMAC-SHA256(key, previous_event_hash + event_json)
-//
-// GAP-05 (Issue #1505): unlike calculateEventHash (plain SHA256), forging a
-// valid hash without the key is computationally infeasible even for an
-// attacker with full read/write access to the database — closing the gap
-// where a DB-privileged attacker could tamper with an event and recompute a
-// self-consistent unkeyed SHA256 chain.
-func calculateEventHashHMAC(key []byte, previousHash string, event *AuditEvent) (string, error) {
-	eventForHashing := PrepareEventForHashing(event)
-
-	eventJSON, err := json.Marshal(eventForHashing)
-	if err != nil {
-		return "", fmt.Errorf("failed to marshal event for hashing: %w", err)
-	}
-
-	mac := hmac.New(sha256.New, key)
-	mac.Write([]byte(previousHash))
-	mac.Write(eventJSON)
-	hashBytes := mac.Sum(nil)
-
-	return hex.EncodeToString(hashBytes), nil
-}
-
-// hashEvent computes the hash chain link for event using this repository's
-// configured algorithm (HMAC-SHA256 when an HMAC key is set, else the legacy
-// unkeyed SHA256) and stamps event.HashAlgorithm accordingly. Used by both
-// Create and CreateBatch so the algorithm decision lives in exactly one place.
-func (r *AuditEventsRepository) hashEvent(previousHash string, event *AuditEvent) (string, error) {
-	if len(r.hmacKey) > 0 {
-		event.HashAlgorithm = HashAlgorithmHMACSHA256
-		return calculateEventHashHMAC(r.hmacKey, previousHash, event)
-	}
-	event.HashAlgorithm = HashAlgorithmSHA256Unkeyed
-	return calculateEventHash(previousHash, event)
-}
-
-// CalculateHashForVerification computes the expected hash for verification,
-// honoring the event's own HashAlgorithm (set at write time). GAP-05 (Issue
-// #1505): supports both the legacy unkeyed SHA256 algorithm and the newer
-// keyed HMAC-SHA256 algorithm within the same chain, since an HMAC key
-// rollout does not retroactively re-hash pre-existing events.
-//
-// hmacKey may be nil when the caller has no HMAC key configured. Verifying a
-// hmac-sha256 event without a key returns an error rather than silently
-// falling back to the unkeyed formula, which would always mismatch and
-// misleadingly report "tampered" for events that are actually intact.
-// Shared by repository.Export and the /api/v1/audit/verify-chain handler so
-// both verification paths apply identical logic.
-func CalculateHashForVerification(hmacKey []byte, previousHash string, event *AuditEvent) (string, error) {
-	switch event.HashAlgorithm {
-	case HashAlgorithmHMACSHA256:
-		if len(hmacKey) == 0 {
-			return "", fmt.Errorf("event %s uses hash_algorithm=%s but no HMAC key is configured for verification", event.EventID, HashAlgorithmHMACSHA256)
-		}
-		return calculateEventHashHMAC(hmacKey, previousHash, event)
-	case HashAlgorithmSHA256Unkeyed, "":
-		// Empty HashAlgorithm covers rows read before the hash_algorithm column
-		// existed (pre-GAP-05 code paths / tests using bare AuditEvent structs).
-		return calculateEventHash(previousHash, event)
-	default:
-		return "", fmt.Errorf("event %s has unrecognized hash_algorithm %q", event.EventID, event.HashAlgorithm)
-	}
-}
-
-// getPreviousEventHash retrieves the hash of the most recent event for a given correlation_id
-// Returns empty string if no previous event exists (first event in chain)
-// Uses advisory lock to prevent race conditions during concurrent inserts
-func (r *AuditEventsRepository) getPreviousEventHash(ctx context.Context, tx *sql.Tx, correlationID string) (string, error) {
-	// Step 1: Acquire advisory lock for this correlation_id (prevents race conditions)
-	// Uses PostgreSQL function audit_event_lock_id() from migration 023
-	_, err := tx.ExecContext(ctx, "SELECT pg_advisory_xact_lock(audit_event_lock_id($1))", correlationID)
-	if err != nil {
-		return "", fmt.Errorf("failed to acquire advisory lock: %w", err)
-	}
-
-	// Step 2: Query last event hash for this correlation_id
-	var previousHash sql.NullString
-	query := `
-		SELECT event_hash
-		FROM audit_events
-		WHERE correlation_id = $1
-		  AND event_hash IS NOT NULL
-		ORDER BY event_timestamp DESC, event_id DESC
-		LIMIT 1
-	`
-
-	err = tx.QueryRowContext(ctx, query, correlationID).Scan(&previousHash)
-	if errors.Is(err, sql.ErrNoRows) {
-		// First event in chain - no previous hash (return empty string)
-		return "", nil
-	}
-	if err != nil {
-		return "", fmt.Errorf("failed to query previous event hash: %w", err)
-	}
-
-	return previousHash.String, nil
-}
-
-// ========================================
-// END GAP #9 HASH CHAIN FUNCTIONS
-// ========================================
-
-// Create inserts a new audit event into the unified audit_events table with hash chain
-// Returns the created event with event_id and created_at populated
-// SOC2 Gap #9: Implements blockchain-style hash chain for tamper detection
-func (r *AuditEventsRepository) Create(ctx context.Context, event *AuditEvent) (*AuditEvent, error) {
-	// Generate UUID if not provided
-	if event.EventID == uuid.Nil {
-		event.EventID = uuid.New()
-	}
-
-	// Set event_timestamp if not provided
-	if event.EventTimestamp.IsZero() {
-		event.EventTimestamp = time.Now().UTC()
-	}
-
-	// Force UTC normalization before hash calculation.
-	// Export/Verify convert to UTC when recomputing hashes (audit_export.go:195,
-	// audit_verify_chain_handler.go:183). Without this, a caller-provided timestamp
-	// with a timezone offset (e.g., "-05:00") would produce a different JSON
-	// representation than the UTC version ("Z"), breaking hash verification.
-	event.EventTimestamp = event.EventTimestamp.UTC()
-
-	// CRITICAL: Truncate to microsecond precision to match PostgreSQL timestamptz
-	// PostgreSQL stores timestamps with microsecond precision (6 decimal places).
-	// Go's time.Time has nanosecond precision (9 decimal places).
-	// If we calculate the hash with nanosecond precision but PostgreSQL stores microseconds,
-	// verification will fail because the JSON will have different timestamps:
-	//   Creation:     "2026-01-24T10:30:45.123456789Z" (9 digits)
-	//   Verification: "2026-01-24T10:30:45.123456Z"    (6 digits)
-	// This causes different JSON → different hash → broken chain.
-	event.EventTimestamp = event.EventTimestamp.Truncate(time.Microsecond)
-
-	// Set event_date from event_timestamp (for partitioning)
-	event.EventDate = DateOnly(time.Date(
-		event.EventTimestamp.Year(),
-		event.EventTimestamp.Month(),
-		event.EventTimestamp.Day(),
-		0, 0, 0, 0, time.UTC,
-	))
-
-	// Marshal event_data to JSONB
-	// CRITICAL: Normalize event_data through JSON round-trip to ensure hash consistency
-	// PostgreSQL JSONB stores and returns JSON with normalized number types (all numbers become float64 in Go).
-	// To ensure the hash calculated during INSERT matches the hash during VERIFICATION,
-	// we must normalize event_data by round-tripping through JSON before hashing.
-	eventDataJSON, err := json.Marshal(event.EventData)
-	if err != nil {
-		return nil, fmt.Errorf("failed to marshal event_data: %w", err)
-	}
-
-	// Normalize: unmarshal and remarshal to match PostgreSQL's representation
-	var normalizedEventData map[string]interface{}
-	if len(eventDataJSON) > 0 && string(eventDataJSON) != "null" {
-		if err := json.Unmarshal(eventDataJSON, &normalizedEventData); err != nil {
-			return nil, fmt.Errorf("failed to normalize event_data: %w", err)
-		}
-		event.EventData = normalizedEventData // Replace with normalized version for hash calculation
-		// Re-marshal for database storage
-		eventDataJSON, err = json.Marshal(normalizedEventData)
-		if err != nil {
-			return nil, fmt.Errorf("failed to remarshal normalized event_data: %w", err)
-		}
-	}
-
-	// Prepare SQL statement (28 columns - added event_version, parent_event_date for FK constraint)
-	// Note: event_date MUST be explicitly set for partitioned tables (triggers don't work on partitions)
-	// Calculate event_date from event_timestamp
-	eventDate := event.EventTimestamp.Truncate(24 * time.Hour)
-
-	// Set default version if not specified (ADR-034: current version is "1.0")
-	version := event.Version
-	if version == "" {
-		version = "1.0"
-	}
-
-	// CRITICAL: Set default retention days BEFORE hash calculation
-	// This ensures the hash includes the correct retention_days value (2555)
-	// instead of 0, matching what will be read back from the DB during verification.
-	if event.RetentionDays == 0 {
-		event.RetentionDays = 2555
-	}
-
-	// ========================================
-	// SOC2 Gap #9: Hash Chain Integration
-	// ========================================
-	// Wrap the transactional portion in a retry loop so that transient
-	// PostgreSQL deadlocks (40P01) are retried transparently. The hash
-	// fields are recalculated on each attempt because they depend on the
-	// current chain head, which may change between retries.
-	err = txretry.WithSerializableRetry(ctx, 3, func() error {
-		event.EventHash = ""
-		event.PreviousEventHash = ""
-
-		tx, txErr := r.db.BeginTx(ctx, nil)
-		if txErr != nil {
-			return fmt.Errorf("failed to begin transaction: %w", txErr)
-		}
-		defer func() {
-			if txErr != nil {
-				_ = tx.Rollback()
-			}
-		}()
-
-		previousHash, txErr := r.getPreviousEventHash(ctx, tx, event.CorrelationID)
-		if txErr != nil {
-			return fmt.Errorf("failed to get previous event hash: %w", txErr)
-		}
-
-		eventHash, txErr := r.hashEvent(previousHash, event)
-		if txErr != nil {
-			return fmt.Errorf("failed to calculate event hash: %w", txErr)
-		}
-
-		event.EventHash = eventHash
-		event.PreviousEventHash = previousHash
-
-		query := `
-			INSERT INTO audit_events (
-				event_id, event_version, event_timestamp, event_date, event_type,
-				event_category, event_action, event_outcome,
-				correlation_id, parent_event_id, parent_event_date,
-				resource_type, resource_id, namespace, cluster_name,
-				actor_id, actor_type, actor_ip,
-				severity, duration_ms, error_code, error_message,
-				retention_days, is_sensitive, event_data,
-				event_hash, previous_event_hash, hash_algorithm,
-				legal_hold, legal_hold_reason, legal_hold_placed_by, legal_hold_placed_at
-			) VALUES (
-				$1, $2, $3, $4, $5,
-				$6, $7, $8,
-				$9, $10, $11,
-				$12, $13, $14, $15,
-				$16, $17, $18,
-				$19, $20, $21, $22,
-				$23, $24, $25,
-				$26, $27, $28,
-				$29, $30, $31, $32
-			)
-			RETURNING event_timestamp
-		`
-
-		parentEventID := sqlutil.ToNullUUID(event.ParentEventID)
-		parentEventDate := sqlutil.ToNullTime(event.ParentEventDate)
-		namespace := sqlutil.ToNullStringValue(event.ResourceNamespace)
-		clusterName := sqlutil.ToNullStringValue(event.ClusterID)
-		errorCode := sqlutil.ToNullStringValue(event.ErrorCode)
-		errorMessage := sqlutil.ToNullStringValue(event.ErrorMessage)
-		severity := sqlutil.ToNullStringValue(event.Severity)
-		actorIP := sqlutil.ToNullStringValue(event.ActorIP)
-
-		var durationMs sql.NullInt32
-		if event.DurationMs != 0 {
-			durationMs = sql.NullInt32{Int32: int32(event.DurationMs), Valid: true}
-		}
-		retentionDays := event.RetentionDays
-
-		var ignoredTimestamp time.Time
-		txErr = tx.QueryRowContext(ctx, query,
-			event.EventID,
-			version,
-			event.EventTimestamp,
-			eventDate,
-			event.EventType,
-			event.EventCategory,
-			event.EventAction,
-			event.EventOutcome,
-			event.CorrelationID,
-			parentEventID,
-			parentEventDate,
-			event.ResourceType,
-			event.ResourceID,
-			namespace,
-			clusterName,
-			event.ActorID,
-			event.ActorType,
-			actorIP,
-			severity,
-			durationMs,
-			errorCode,
-			errorMessage,
-			retentionDays,
-			event.IsSensitive,
-			eventDataJSON,
-			event.EventHash,
-			event.PreviousEventHash,
-			event.HashAlgorithm,
-			event.LegalHold,
-			sqlutil.ToNullStringValue(event.LegalHoldReason),
-			sqlutil.ToNullStringValue(event.LegalHoldPlacedBy),
-			sqlutil.ToNullTime(event.LegalHoldPlacedAt),
-		).Scan(&ignoredTimestamp)
-
-		if txErr != nil {
-			return fmt.Errorf("failed to insert audit event: %w", txErr)
-		}
-
-		if txErr = tx.Commit(); txErr != nil {
-			return fmt.Errorf("failed to commit transaction: %w", txErr)
-		}
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	r.logger.V(1).Info("Audit event created with hash chain",
-		"event_id", event.EventID.String(),
-		"event_type", event.EventType,
-		"event_category", event.EventCategory,
-		"correlation_id", event.CorrelationID,
-		"event_hash", event.EventHash[:16]+"...", // Log first 16 chars for debugging
-		"has_previous_hash", event.PreviousEventHash != "",
-	)
-
-	return event, nil
-}
-
-// CreateBatch inserts multiple audit events in a single transaction
-// DD-AUDIT-002: StoreBatch interface for batch audit event storage
-// SortedCorrelationIDs returns the keys of a map in lexicographic order.
-// Issue #667 / BR-STORAGE-040: Deterministic ordering prevents advisory lock
-// deadlocks when multiple concurrent transactions lock overlapping correlation IDs.
-func SortedCorrelationIDs[V any](m map[string]V) []string {
-	keys := make([]string, 0, len(m))
-	for k := range m {
-		keys = append(keys, k)
-	}
-	sort.Strings(keys)
-	return keys
-}
-
-// BR-AUDIT-001: Complete audit trail with no data loss
-// Uses a single transaction for atomic batch insert (all succeed or all fail).
-// Wraps the transaction in a retry loop so that transient PostgreSQL deadlocks
-// (40P01) are retried transparently.
-func (r *AuditEventsRepository) CreateBatch(ctx context.Context, events []*AuditEvent) ([]*AuditEvent, error) {
-	if len(events) == 0 {
-		return nil, fmt.Errorf("batch cannot be empty")
-	}
-
-	type indexedEvent struct {
-		originalIndex int
-		event         *AuditEvent
-	}
-
-	// Normalize fields that are idempotent across retries (UUID, timestamp,
-	// event_data JSON round-trip, retention defaults). These only need to run
-	// once regardless of how many retry attempts occur.
-	for _, event := range events {
-		if event.EventID == uuid.Nil {
-			event.EventID = uuid.New()
-		}
-		if event.EventTimestamp.IsZero() {
-			event.EventTimestamp = time.Now().UTC()
-		}
-		event.EventTimestamp = event.EventTimestamp.UTC().Truncate(time.Microsecond)
-		event.EventDate = DateOnly(event.EventTimestamp.Truncate(24 * time.Hour))
-
-		eventDataJSON, marshalErr := json.Marshal(event.EventData)
-		if marshalErr != nil {
-			return nil, fmt.Errorf("failed to marshal event_data for event %s: %w", event.EventID, marshalErr)
-		}
-		var normalizedEventData map[string]interface{}
-		if len(eventDataJSON) > 0 && string(eventDataJSON) != "null" {
-			if unmarshalErr := json.Unmarshal(eventDataJSON, &normalizedEventData); unmarshalErr != nil {
-				return nil, fmt.Errorf("failed to normalize event_data for event %s: %w", event.EventID, unmarshalErr)
-			}
-			event.EventData = normalizedEventData
-		}
-		if event.Version == "" {
-			event.Version = "1.0"
-		}
-		if event.RetentionDays == 0 {
-			event.RetentionDays = 2555
-		}
-	}
-
-	// Group events by correlation_id (stable across retries).
-	eventsByCorrelation := make(map[string][]indexedEvent)
-	for i, event := range events {
-		eventsByCorrelation[event.CorrelationID] = append(
-			eventsByCorrelation[event.CorrelationID],
-			indexedEvent{originalIndex: i, event: event},
-		)
-	}
-
-	// Issue #667 / BR-STORAGE-040: sorted order prevents advisory-lock deadlocks.
-	sortedCorrIDs := SortedCorrelationIDs(eventsByCorrelation)
-
-	var createdEvents []*AuditEvent
-
-	err := txretry.WithSerializableRetry(ctx, 3, func() error {
-		// Reset hash fields so they are recalculated from the current chain head.
-		for _, event := range events {
-			event.EventHash = ""
-			event.PreviousEventHash = ""
-			event.HashAlgorithm = ""
-		}
-
-		tx, txErr := r.db.BeginTx(ctx, nil)
-		if txErr != nil {
-			return fmt.Errorf("failed to begin transaction: %w", txErr)
-		}
-		defer func() {
-			if txErr != nil {
-				_ = tx.Rollback()
-			}
-		}()
-
-		query := `
-			INSERT INTO audit_events (
-				event_id, event_version, event_timestamp, event_date, event_type,
-				event_category, event_action, event_outcome,
-				correlation_id, parent_event_id, parent_event_date,
-				resource_type, resource_id, namespace, cluster_name,
-				actor_id, actor_type, actor_ip,
-				severity, duration_ms, error_code, error_message,
-				retention_days, is_sensitive, event_data,
-				event_hash, previous_event_hash, hash_algorithm,
-				legal_hold, legal_hold_reason, legal_hold_placed_by, legal_hold_placed_at
-			) VALUES (
-				$1, $2, $3, $4, $5,
-				$6, $7, $8,
-				$9, $10, $11,
-				$12, $13, $14, $15,
-				$16, $17, $18,
-				$19, $20, $21, $22,
-				$23, $24, $25,
-				$26, $27, $28,
-				$29, $30, $31, $32
-			)
-			RETURNING event_timestamp
-		`
-
-		stmt, txErr := tx.PrepareContext(ctx, query)
-		if txErr != nil {
-			return fmt.Errorf("failed to prepare statement: %w", txErr)
-		}
-		defer func() { _ = stmt.Close() }()
-
-		result := make([]*AuditEvent, len(events))
-		lastHashByCorrelation := make(map[string]string)
-
-		for _, correlationID := range sortedCorrIDs {
-			correlationEvents := eventsByCorrelation[correlationID]
-
-			previousHash, hashErr := r.getPreviousEventHash(ctx, tx, correlationID)
-			if hashErr != nil {
-				txErr = fmt.Errorf("failed to get previous event hash for correlation_id %s: %w", correlationID, hashErr)
-				return txErr
-			}
-			lastHashByCorrelation[correlationID] = previousHash
-
-			for _, ie := range correlationEvents {
-				event := ie.event
-
-				eventDataJSON, marshalErr := json.Marshal(event.EventData)
-				if marshalErr != nil {
-					txErr = fmt.Errorf("failed to marshal event_data for event %s in batch insert: %w", event.EventID, marshalErr)
-					return txErr
-				}
-				eventDate := event.EventTimestamp.Truncate(24 * time.Hour)
-
-				previousHash := lastHashByCorrelation[event.CorrelationID]
-				eventHash, hashErr := r.hashEvent(previousHash, event)
-				if hashErr != nil {
-					txErr = fmt.Errorf("failed to calculate event hash for event %s: %w", event.EventID, hashErr)
-					return txErr
-				}
-
-				event.EventHash = eventHash
-				event.PreviousEventHash = previousHash
-				lastHashByCorrelation[event.CorrelationID] = eventHash
-
-				parentEventID := sqlutil.ToNullUUID(event.ParentEventID)
-				parentEventDate := sqlutil.ToNullTime(event.ParentEventDate)
-				namespace := sqlutil.ToNullStringValue(event.ResourceNamespace)
-				clusterName := sqlutil.ToNullStringValue(event.ClusterID)
-				errorCode := sqlutil.ToNullStringValue(event.ErrorCode)
-				errorMessage := sqlutil.ToNullStringValue(event.ErrorMessage)
-				severity := sqlutil.ToNullStringValue(event.Severity)
-				actorIP := sqlutil.ToNullStringValue(event.ActorIP)
-
-				var durationMs sql.NullInt32
-				if event.DurationMs != 0 {
-					durationMs = sql.NullInt32{Int32: int32(event.DurationMs), Valid: true}
-				}
-
-				var returnedTimestamp time.Time
-				txErr = stmt.QueryRowContext(ctx,
-					event.EventID,
-					event.Version,
-					event.EventTimestamp,
-					eventDate,
-					event.EventType,
-					event.EventCategory,
-					event.EventAction,
-					event.EventOutcome,
-					event.CorrelationID,
-					parentEventID,
-					parentEventDate,
-					event.ResourceType,
-					event.ResourceID,
-					namespace,
-					clusterName,
-					event.ActorID,
-					event.ActorType,
-					actorIP,
-					severity,
-					durationMs,
-					errorCode,
-					errorMessage,
-					event.RetentionDays,
-					event.IsSensitive,
-					eventDataJSON,
-					event.EventHash,
-					event.PreviousEventHash,
-					event.HashAlgorithm,
-					event.LegalHold,
-					sqlutil.ToNullStringValue(event.LegalHoldReason),
-					sqlutil.ToNullStringValue(event.LegalHoldPlacedBy),
-					sqlutil.ToNullTime(event.LegalHoldPlacedAt),
-				).Scan(&returnedTimestamp)
-
-				if txErr != nil {
-					return fmt.Errorf("failed to insert event %s: %w", event.EventID, txErr)
-				}
-
-				result[ie.originalIndex] = event
-			}
-		}
-
-		if txErr = tx.Commit(); txErr != nil {
-			return fmt.Errorf("failed to commit batch transaction: %w", txErr)
-		}
-
-		createdEvents = result
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	r.logger.Info("Batch audit events created with hash chains",
-		"count", len(createdEvents),
-		"correlation_ids", len(eventsByCorrelation),
-	)
-
-	return createdEvents, nil
-}
-
-// PaginationMetadata contains pagination information for query results
-// DD-STORAGE-010: Offset-based pagination metadata
-type PaginationMetadata struct {
-	Limit   int  `json:"limit"`
-	Offset  int  `json:"offset"`
-	Total   int  `json:"total"`
-	HasMore bool `json:"has_more"`
-}
-
-// Query retrieves audit events based on filters with pagination
-// DD-STORAGE-010: Query API with offset-based pagination
-// BR-STORAGE-021: REST API Read Endpoints
-// BR-STORAGE-022: Query Filtering
-// BR-STORAGE-023: Pagination Support
-func (r *AuditEventsRepository) Query(ctx context.Context, querySQL string, countSQL string, args []interface{}) ([]*AuditEvent, *PaginationMetadata, error) {
-	// Execute count query for pagination metadata
-	var total int
-	// Safely exclude limit and offset from count query args
-	// Fix: Prevent panic if args has fewer than 2 elements
-	countArgs := args
-	if len(args) >= 2 {
-		countArgs = args[:len(args)-2] // Exclude limit and offset for count query
-	}
-	err := r.db.QueryRowContext(ctx, countSQL, countArgs...).Scan(&total)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to count audit events: %w", err)
-	}
-
-	// Execute main query
-	rows, err := r.db.QueryContext(ctx, querySQL, args...)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to query audit events: %w", err)
-	}
-	defer func() { _ = rows.Close() }()
-
-	// Parse results
-	events := make([]*AuditEvent, 0)
-	for rows.Next() {
-		event := &AuditEvent{}
-		var eventDataJSON []byte
-		var parentEventID sql.NullString
-		var actorID, actorType, actorIP, resourceType, resourceID sql.NullString
-		var severity, namespace, clusterName sql.NullString
-		var errorCode, errorMessage sql.NullString // DD-TESTING-001: Error fields
-		var durationMs sql.NullInt64               // DD-TESTING-001: Performance tracking (BR-SP-090)
-		var eventHash, previousEventHash sql.NullString
-		var retentionDays sql.NullInt64
-		var isSensitive sql.NullBool
-		var parentEventDate sql.NullTime
-		var legalHold sql.NullBool
-		var legalHoldReason, legalHoldPlacedBy sql.NullString
-		var legalHoldPlacedAt sql.NullTime
-
-		err := rows.Scan(
-			&event.EventID,
-			&event.Version, // event_version from DB (maps to version in OpenAPI)
-			&event.EventType,
-			&event.EventCategory, // ADR-034
-			&event.EventAction,   // ADR-034
-			&event.CorrelationID,
-			&event.EventTimestamp,
-			&event.EventOutcome, // ADR-034
-			&severity,
-			&resourceType,
-			&resourceID,
-			&actorType,
-			&actorID,
-			&actorIP,
-			&parentEventID,
-			&eventDataJSON,
-			&event.EventDate,
-			&namespace,
-			&clusterName,
-			&durationMs,   // DD-TESTING-001: Added for top-level field validation
-			&errorCode,    // DD-TESTING-001: Added for error validation
-			&errorMessage, // DD-TESTING-001: Added for error validation
-			&eventHash,
-			&previousEventHash,
-			&retentionDays,
-			&isSensitive,
-			&parentEventDate,
-			&legalHold,
-			&legalHoldReason,
-			&legalHoldPlacedBy,
-			&legalHoldPlacedAt,
-		)
-		if err != nil {
-			return nil, nil, fmt.Errorf("failed to scan audit event: %w", err)
-		}
-
-		// Handle NULL fields
-		if severity.Valid {
-			event.Severity = severity.String
-		}
-		if resourceType.Valid {
-			event.ResourceType = resourceType.String
-		}
-		if resourceID.Valid {
-			event.ResourceID = resourceID.String
-		}
-		if actorType.Valid {
-			event.ActorType = actorType.String
-		}
-		if actorID.Valid {
-			event.ActorID = actorID.String
-		}
-		if actorIP.Valid {
-			event.ActorIP = actorIP.String
-		}
-		if parentEventID.Valid {
-			parentUUID, err := uuid.Parse(parentEventID.String)
-			if err == nil {
-				event.ParentEventID = &parentUUID
-			}
-		}
-		if namespace.Valid {
-			event.ResourceNamespace = namespace.String
-		}
-		if clusterName.Valid {
-			event.ClusterID = clusterName.String
-		}
-		// DD-TESTING-001: Handle optional fields for comprehensive audit validation
-		if durationMs.Valid {
-			event.DurationMs = int(durationMs.Int64) // BR-SP-090: Performance tracking
-		}
-		if errorCode.Valid {
-			event.ErrorCode = errorCode.String
-		}
-		if errorMessage.Valid {
-			event.ErrorMessage = errorMessage.String
-		}
-		if eventHash.Valid {
-			event.EventHash = eventHash.String
-		}
-		if previousEventHash.Valid {
-			event.PreviousEventHash = previousEventHash.String
-		}
-		if retentionDays.Valid {
-			event.RetentionDays = int(retentionDays.Int64)
-		}
-		if isSensitive.Valid {
-			event.IsSensitive = isSensitive.Bool
-		}
-		if parentEventDate.Valid {
-			ts := parentEventDate.Time
-			event.ParentEventDate = &ts
-		}
-		if legalHold.Valid {
-			event.LegalHold = legalHold.Bool
-		}
-		if legalHoldReason.Valid {
-			event.LegalHoldReason = legalHoldReason.String
-		}
-		if legalHoldPlacedBy.Valid {
-			event.LegalHoldPlacedBy = legalHoldPlacedBy.String
-		}
-		if legalHoldPlacedAt.Valid {
-			ts := legalHoldPlacedAt.Time
-			event.LegalHoldPlacedAt = &ts
-		}
-
-		// Unmarshal event_data JSONB
-		if len(eventDataJSON) > 0 {
-			if err := json.Unmarshal(eventDataJSON, &event.EventData); err != nil {
-				return nil, nil, fmt.Errorf("failed to unmarshal event_data: %w", err)
-			}
-		}
-
-		events = append(events, event)
-	}
-
-	if err := rows.Err(); err != nil {
-		return nil, nil, fmt.Errorf("error iterating audit events: %w", err)
-	}
-
-	// Build pagination metadata
-	// Safely extract limit and offset from args (default to 0 if not present)
-	limit := 0
-	offset := 0
-	if len(args) >= 2 {
-		limit = int(args[len(args)-2].(int))
-		offset = int(args[len(args)-1].(int))
-	} else if len(args) == 1 {
-		limit = int(args[0].(int))
-	}
-	pagination := &PaginationMetadata{
-		Limit:   limit,
-		Offset:  offset,
-		Total:   total,
-		HasMore: offset+len(events) < total,
-	}
-
-	r.logger.V(1).Info("Audit events queried",
-		"count", len(events),
-		"total", total,
-		"limit", limit,
-		"offset", offset,
-	)
-
-	return events, pagination, nil
-}
-
-// HealthCheck verifies database connectivity
-func (r *AuditEventsRepository) HealthCheck(ctx context.Context) error {
-	if err := r.db.PingContext(ctx); err != nil {
-		return fmt.Errorf("database health check failed: %w", err)
-	}
-	return nil
 }
